Reconstruct original-to-instrumented mapping outside of transformation
As a precondition to migrating the configuration cache instrumentation
to ArtifactTransform, we must be able to build the classloader from the
ArtifactTransform output. The latter is limited compared to the ad-hoc
transformation we have, and cannot carry necessary
original-to-instrumented mappings as a TransformedClassPath object. The
only available option is a list of files.

This CL makes CachedClasspathTransformer more similar to the
ArtifactTransform in this regard. It now outputs a list of instrumented
JAR-original JAR pairs. This list is converted to TransformedClassPath
quite fast to avoid further changes downstream, as I believe that only
a small, known subset of classloaders should deal with the
instrumentation.